Initial commit of starting point for the new nfl-pickem
This commit is the beginning of the feature branch in
nfl-pickem-legacy (github.com/ameske/nfl-pickem-legacy) where the
transition to a RESTful API backend started. While the transition reused
bits and pieces of the legacy nfl-pickem's code, it really is a completely
different application that is not backwards compatible with legacy. Thus,
I didn't want to bring over all of the git history associated
with the old version of the application.