Implement the beginning of a  CLI management application
At the moment there aren't a lot of administration endpoints in the
backend daemon, so there needs to be away to do things such as edit
games, picks, or schedules. A CLI app like git or docker is a good
method of doing so. This sets up the structure using spf13/cobra and
implements a schedule download command/subcommand to get things started.