sqlite3: correct bug where null selection picks weren't being returned
The previous DDL for the picks table did not have the selection field as
a FK to the teams table, thus it was always set to 0, 1, or 2. When
updating the app for the new structure I was testing against historical
data where picks were made and missed the fact that when picks haven't
been made the current SQL wouldn't return the unselected picks.

This splits the picks off into Selected and Unselected picks. The
interface implementation now has a UserPicks (for specific user
filtering) and Picks (for all of them). Those implementations in
sqlite3 call Selected and Unselected to get the full pick set. The
Selected and Unselected functions remain exported for now.