Add the ability to build a docker image for the nfl-pickem app
Things have gotten to the point where it is neccessary to test the app
as if it was deployed. I've always been interested in the mechanics of
docker and exploring how useful it is for development, so this scratched
an itch.

This commit adds the ability to build a docker image that uses the Caddy
webserver to serve the NFL pickem HTML/JS and the backend RESTful API.
It uses a multi-stage build to first compile the nfl-pickem app in a
Debian container (as the base image I used is Debian from the Docker
hub). The Dockerfile is set up to invalidate the docker cache if you
change one of the source files, so it is possible to develop and
iterate by rebuilidng the container.

I think this is ultimately going to be how I deploy the app in
production, as this has been a very pleasant experience.